Stripped-down implementation showing routing but without any default values.
