+++ src/Restival.Api.Common/Resources/Greeting.cs
@@ -2,8 +2,7 @@ using System;
 
 namespace Restival.Api.Common.Resources {
     public class Greeting {
-        public Greeting()
-            : this("World") { }
+        public Greeting() { }
 
         public Greeting(string name) {
             Message = String.Format("Hello, {0}!", name);
