+++ src/Restival.Api.Nancy/HelloModule.cs
@@ -5,7 +5,7 @@ namespace Restival.Api.Nancy {
     public class HelloModule : NancyModule {
         public HelloModule() {
             Get["/hello"] = _ => new Greeting();
-            Get["/hello/{name}"] = _ => new Greeting(_.name);
+            Get["/hello/{name}"] = parameters => new Greeting(parameters.name);
         }
     }
 }
