+++ src/Restival.Api.OpenRasta/Global.asax.cs
@@ -1,25 +1,9 @@
 ﻿using System.Web;
-using OpenRasta.Configuration;
-using Restival.Api.Common.Resources;
-using Restival.Api.OpenRasta.Handlers;
 
 namespace Restival.Api.OpenRasta {
     // Note: For instructions on enabling IIS6 or IIS7 classic mode, 
     // visit http://go.microsoft.com/?LinkId=9394801
     public class MvcApplication : HttpApplication {
         protected void Application_Start() { }
-    }
-
-    public class Configuration : IConfigurationSource {
-        public void Configure() {
-            using (OpenRastaConfiguration.Manual) {
-                ResourceSpace.Has.ResourcesOfType<Greeting>()
-                    .AtUri("/hello")
-                    .And.AtUri("/hello/{name}")
-                    .HandledBy<HelloHandler>()
-                    .AsXmlDataContract()
-                    .And.AsJsonDataContract();
-            }
-        }
     }
 }
