+++ src/Restival.Api.OpenRasta/Handlers/HelloHandler.cs
@@ -2,10 +2,6 @@
 
 namespace Restival.Api.OpenRasta.Handlers {
     public class HelloHandler {
-        public object Get() {
-            return (new Greeting());
-        }
-
         public object Get(string name) {
             return (new Greeting(name));
         }
