+++ src/Restival.Api.ServiceStack/Restival.Api.ServiceStack.csproj
@@ -132,6 +132,7 @@
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\HelloRequestDto.cs" />
+    <Compile Include="Services\HelloResponse.cs" />
     <Compile Include="Services\HelloService.cs" />
   </ItemGroup>
   <ItemGroup>
