+++ src/Restival.Api.ServiceStack/Services/HelloRequestDto.cs
@@ -3,7 +3,7 @@
 namespace Restival.Api.ServiceStack.Services {
     [Route("/hello")]
     [Route("/hello/{name}")]
-    public class HelloRequestDto {
+    public class Hello {
         public string Name { get; set; }
     }
 }
