+++ src/Restival.Api.ServiceStack/Services/HelloResponse.cs
@@ -1,0 +1,5 @@
+﻿namespace Restival.Api.ServiceStack.Services {
+    public class HelloResponse {
+        public string Message { get; set; }
+    }
+}
