+++ src/Restival.Api.ServiceStack/Services/HelloService.cs
@@ -3,9 +3,10 @@ using ServiceStack;
 
 namespace Restival.Api.ServiceStack.Services {
     public class HelloService : Service {
-        public Greeting Get(HelloRequestDto dto) {
-            if (dto.Name == null) return (new Greeting());
-            return (new Greeting(dto.Name));
+        public HelloResponse Any(Hello dto) {
+            var greeting = new Greeting(dto.Name);
+            var response = new HelloResponse() { Message = greeting.Message };
+            return (response);
         }
     }
 }
