+++ src/Restival.Api.WebApi/Controllers/HelloController.cs
@@ -3,10 +3,6 @@ using Restival.Api.Common.Resources;
 
 namespace Restival.Api.WebApi.Controllers {
     public class HelloController : ApiController {
-        public Greeting Get() {
-            return (new Greeting());
-        }
-
         public Greeting Get(string name) {
             return (new Greeting(name));
         }
